Replace TRGBEImage class with TRGBFloatImage class
This means that the float image in *memory* is stored as 3 x Single, so it takes more space in memory than 4 x Byte. However, memory space is not so crucial, and this is much more natural --- you can perform many operations fast, without the need to decode RGBE<->3Single. Also, RGBE is not widely known (OpenEXR is probably the float format of choice now), so for the future --- generic float format is more natural than a specialized RGBE format.

git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/rayhunter@6223 afe8dcc5-9094-451b-8773-5f495042a737